% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant.R
\name{discriminant}
\alias{discriminant}
\alias{discriminant.default}
\alias{discriminant.formula}
\title{Discriminant meta-features}
\usage{
discriminant(...)

\method{discriminant}{default}(x, y, features = "all", ...)

\method{discriminant}{formula}(formula, data, features = "all", ...)
}
\arguments{
\item{...}{Not used.}

\item{x}{A data.frame contained only the input attributes.}

\item{y}{A factor response vector with one label for each row/component of x.}

\item{features}{A list of features names or \code{"all"} to include all them.}

\item{formula}{A formula to define the class column.}

\item{data}{A data.frame dataset contained the input attributes and class
The details section describes the valid values for this group.}
}
\value{
A list named by the requested meta-features.
}
\description{
Discriminant meta-features are computed using the discriminant analysis
algorithm. It is computed using just the numerical attributes so a
binarization on the data is required.
}
\details{
The following features are allowed for this method:
 \describe{
   \item{"cancor"}{Represents the first canonical discriminant correlations
      of the numeric attributes in the dataset.}
   \item{"center.of.gravity"}{Represents the distance between minority and
      majority classes in the dataset.}
   \item{"discfct"}{Represents the number of discriminant functions
      normalized by the number of classes"}
   \item{"cancor.fract"}{Represents a measure of collinearity of the class
      means. It is computed using the squares of the canonical correlations
      instead of the eigenvalues.}
   \item{"eigen.fract"}{Represents the relative importance of the largest
      eigenvalue of the attribute covariance matrix computed from the
      numeric attributes in the dataset.}
   \item{"max.eigenvalue"}{Represents the maximum eigenvalue.}
   \item{"min.eigenvalue"}{Represents the minimum eigenvalue.}
   \item{"sdratio"}{Represents the test statistic for homogeneity of
      covariances. It uses the Box's M test and it is is strictly greater
      than unity if the covariances differ, and is equal to unity if and only
      if the M-statistic is zero.}
 }

 The categorical attributes is replaced by binaries attributes.
}
\examples{
## Extract all metafeatures
discriminant(Species ~ ., iris)

## Extract some metafeatures
discriminant(iris[1:4], iris[5], c("cancor", "cancor.fract"))
}
\references{
Donald Michie, David J. Spiegelhalter, Charles C. Taylor, and John Campbell. 
 Machine Learning, Neural and Statistical Classification, volume 37. Ellis 
 Horwood Upper Saddle River, 1994.

 Ciro Castiello, Giovanna Castellano, and Anna Maria Fanelli. Meta-data: 
 Characterization of input features for meta-learning. In 2nd International 
 Conference on Modeling Decisions for Artificial Intelligence (MDAI), 
 pages 457 - 468, 2005.

 Guido Lindner and Rudi Studer. AST: Support for algorithm selection with a 
 CBR approach. In European Conference on Principles of Data Mining and 
 Knowledge Discovery (PKDD), pages 418 - 423, 1999.

 Shawkat Ali and Kate A. Smith. On learning algorithm selection for 
 classification. Applied Soft Computing, 6(2):119 - 138, 2006.
}
\seealso{
Other meta-features: \code{\link{general}},
  \code{\link{infotheo}}, \code{\link{landmarking}},
  \code{\link{model.based}}, \code{\link{statistical}}
}
