% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept.R
\name{ls.concept}
\alias{ls.concept}
\title{List the best concept meta-features}
\usage{
ls.concept()
}
\value{
A list of concept meta-features names.
}
\description{
List the best concept meta-features
}
\examples{
ls.concept()
}
