% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminant.R
\name{ls.discriminant}
\alias{ls.discriminant}
\title{List the discriminant meta-features}
\usage{
ls.discriminant()
}
\value{
A list of discriminant meta-features names
}
\description{
List the discriminant meta-features
}
\examples{
ls.discriminant()
}

