% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{ls.general}
\alias{ls.general}
\title{List the general meta-features}
\usage{
ls.general()
}
\value{
A list of general meta-features names
}
\description{
List the general meta-features
}
\examples{
ls.general()
}
