% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infotheo.R
\name{ls.infotheo}
\alias{ls.infotheo}
\title{List the information theoretical meta-features}
\usage{
ls.infotheo()
}
\value{
A list of information theoretical meta-features names
}
\description{
List the information theoretical meta-features
}
\examples{
ls.infotheo()
}
