% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemset.R
\name{ls.itemset}
\alias{ls.itemset}
\title{List the itemset meta-features}
\usage{
ls.itemset()
}
\value{
A list of itemset meta-features names.
}
\description{
List the itemset meta-features
}
\examples{
ls.itemset()
}
