% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{ls.landmarking}
\alias{ls.landmarking}
\title{List the Landmarking meta-features}
\usage{
ls.landmarking()
}
\value{
A list of Landmarking meta-features names.
}
\description{
List the Landmarking meta-features
}
\examples{
ls.landmarking()
}
