% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeatures.R
\name{ls.metafeatures}
\alias{ls.metafeatures}
\title{List the meta-features groups}
\usage{
ls.metafeatures()
}
\value{
A list of meta-features groups
}
\description{
List the meta-features groups
}
\examples{
ls.metafeatures()
}
