% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.based.R
\name{ls.model.based}
\alias{ls.model.based}
\title{List the DT model based meta-features}
\usage{
ls.model.based()
}
\value{
A list of DT model based meta-features names
}
\description{
List the DT model based meta-features
}
\examples{
ls.model.based()
}
