% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative.R
\name{ls.relative}
\alias{ls.relative}
\title{List the relative meta-features}
\usage{
ls.relative()
}
\value{
A list of relative meta-features names.
}
\description{
List the relative meta-features
}
\examples{
ls.relative()
}
