% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{ls.statistical}
\alias{ls.statistical}
\title{List the statistical meta-features}
\usage{
ls.statistical()
}
\value{
A list of statistical meta-features names
}
\description{
List the statistical meta-features
}
\examples{
ls.statistical()
}
