% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{mf.general}
\alias{mf.general}
\alias{mf.general.default}
\alias{mf.general.formula}
\title{General meta-features}
\usage{
mf.general(...)

\method{mf.general}{default}(x, y, features = "all", summary = c("mean",
  "sd"), ...)

\method{mf.general}{formula}(formula, data, features = "all",
  summary = c("mean", "sd"), ...)
}
\arguments{
\item{...}{Not used.}

\item{x}{A data.frame contained only the input attributes.}

\item{y}{A factor response vector with one label for each row/component of x.}

\item{features}{A list of features names or \code{"all"} to include all them.
The supported values are described in the details section. (Default: 
\code{"all"})}

\item{summary}{A list of summarization functions or empty for all values. See
\link{post.processing} method to more information. (Default: 
\code{c("mean", "sd")})}

\item{formula}{A formula to define the class column.}

\item{data}{A data.frame dataset contained the input attributes and class}
}
\value{
A list named by the requested meta-features.
}
\description{
General meta-features include general information related to the dataset. It
is also known as simple measures.
}
\details{
The following features are allowed for this method:
 \describe{
   \item{"attrToInst"}{Ratio of the number of attributes per the number of 
   instances, also known as dimensionality.}
   \item{"catToNum"}{Ratio of the number of categorical attributes per the 
   number of numeric attributes.}
   \item{"defError"}{Default error.}
   \item{"instPerAttr"}{Ratio of the number of instances per the number of 
   attributes.}
   \item{"ntAttr"}{Number of attributes.}
   \item{"nrBin"}{Number of binary attributes.}
   \item{"nrCat"}{Number of categorical attributes.}
   \item{"nrClass"}{Number of classes.}
   \item{"nrInst"}{Number of instances.}
   \item{"nrNum"}{Number of numeric attributes.}
   \item{"numToCat"}{Ratio of the number of numeric attributes per the number
   of categorical attributes.}
   \item{"propBin"}{Proportion of binary attributes.}
   \item{"propCat"}{Proportion of categorical attributes.}
   \item{"propClass"}{Proportion of the classes values (multi-valued).}
   \item{"propNum"}{Proportion of numeric attributes.}
 }
}
\examples{
## Extract all metafeatures
mf.general(Species ~ ., iris)

## Extract some metafeatures
mf.general(iris[1:100, 1:4], iris[1:100, 5], c("defError", "nrClass"))

## Extract all meta-features without summarize prop.class
mf.general(Species ~ ., iris, summary=c())

## Use another summarization functions
mf.general(Species ~ ., iris, summary=c("sd","min","iqr"))
}
\references{
Michie, E. D., Spiegelhalter, D. J., & Taylor, C. C. (1994).
   Machine Learning , Neural and Statistical Classification.
   Technometrics, 37(4), 459.

 Lindner, G., & Studer, R. (1999). AST: Support for Algorithm Selection with
   a CBR Approach. Principles of Data Mining and Knowledge Discovery, 1704,
   418-423.

 Castiello, C., Castellano, G., & Fanelli, A. M. (2005). Meta-data:
   Characterization of Input Features for Meta-learning. In Proceedings of
   the 2nd International Conference on Modeling Decisions for Artificial
   Intelligence (Vol. 3558, pp. 457-468).
}
\seealso{
Other meta-features: \code{\link{mf.discriminant}},
  \code{\link{mf.infotheo}}, \code{\link{mf.landmarking}},
  \code{\link{mf.model.based}},
  \code{\link{mf.statistical}}
}
