% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{mf.statistical}
\alias{mf.statistical}
\alias{mf.statistical.default}
\alias{mf.statistical.formula}
\title{Statistical meta-features}
\usage{
mf.statistical(...)

\method{mf.statistical}{default}(x, y, features = "all", summary = c("mean",
  "sd"), by.class = FALSE, transform = TRUE, ...)

\method{mf.statistical}{formula}(formula, data, features = "all",
  summary = c("mean", "sd"), by.class = FALSE, transform = TRUE, ...)
}
\arguments{
\item{...}{Further arguments passed to the summarization functions.}

\item{x}{A data.frame contained only the input attributes.}

\item{y}{A factor response vector with one label for each row/component of x.}

\item{features}{A list of features names or \code{"all"} to include all them.
The details section describes the valid values for this group.}

\item{summary}{A list of summarization functions or empty for all values. See
\link{post.processing} method to more information. (Default: 
\code{c("mean", "sd")})}

\item{by.class}{A logical value indicating if the meta-features must be
computed for each group of samples belonging to different output classes.
(Default: FALSE)}

\item{transform}{A logical value indicading if the categorical attributes
should be transformed. If \code{FALSE} they will be ignored. (Default: 
\code{TRUE})}

\item{formula}{A formula to define the class column.}

\item{data}{A data.frame dataset contained the input attributes and class
The details section describes the valid values for this group.}
}
\value{
A list named by the requested meta-features.
}
\description{
Statistical meta-features are the standard statistical measures to describe
the numerical properties of a distribution of data. As it requires only
numerical attributes, the categorical data are transformed to numerical.
}
\examples{
## Extract all meta-features
mf.statistical(Species ~ ., iris)

## Extract some meta-features
mf.statistical(iris[1:4], iris[5], c("cor", "nrNorm"))

## Extract all meta-features without summarize the results
mf.statistical(Species ~ ., iris, summary=c())

## Use another summarization function
mf.statistical(Species ~ ., iris, summary=c("min", "median", "max"))

## Extract statistical measures using by.class approach
mf.statistical(Species ~ ., iris, by.class=TRUE)

## Do not transform the data (using only categorical attributes)
mf.statistical(Species ~ ., iris, transform=FALSE)
}
\references{
Castiello, C., Castellano, G., & Fanelli, A. M. (2005). Meta-data:
   Characterization of Input Features for Meta-learning. In Proceedings of
   the 2nd International Conference on Modeling Decisions for Artificial
   Intelligence (Vol. 3558, pp. 457-468).

 Ali, S., & Smith, K. a. (2006). On learning algorithm selection for
   classification. Applied Soft Computing, 6(2), 119-138.
}
\seealso{
Other meta-features: \code{\link{mf.discriminant}},
  \code{\link{mf.general}}, \code{\link{mf.infotheo}},
  \code{\link{mf.landmarking}},
  \code{\link{mf.model.based}}
}
