% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarking.R
\name{mf.landmarking}
\alias{mf.landmarking}
\alias{mf.landmarking.default}
\alias{mf.landmarking.formula}
\title{Landmarking Meta-features}
\usage{
mf.landmarking(...)

\method{mf.landmarking}{default}(x, y, features = "all", summary = c("mean",
  "sd"), map = c("one.vs.all", "one.vs.one"), folds = 10,
  transform.attr = TRUE, ...)

\method{mf.landmarking}{formula}(formula, data, features = "all",
  summary = c("mean", "sd"), map = c("one.vs.all", "one.vs.one"),
  folds = 10, transform.attr = TRUE, ...)
}
\arguments{
\item{...}{Optional arguments to the summary methods.}

\item{x}{A data.frame contained only the input attributes}

\item{y}{A factor response vector with one label for each row/component of x.}

\item{features}{A list of features names or \code{"all"} to include all them.}

\item{summary}{A list of methods to summarize the results as post-processing
functions. See \link{post.processing} method to more information. (Default:
\code{c("mean", "sd")})}

\item{map}{A list of decomposition strategies for multi-class classification
problems. The options are: \code{"one.vs.all"} and \code{"one.vs.one"}
strategy.}

\item{folds}{The number of k equal size subsamples in k-fold
cross-validation.}

\item{transform.attr}{A logical value indicating if the categorical
attributes should be transformed to numerical.}

\item{formula}{A formula to define the class column.}

\item{data}{A data.frame dataset contained the input attributes and class.
The details section describes the valid values for this group.}
}
\value{
Each one of these meta-features generate multiple values (by fold
 and/or binary dataset) and then it is post processed by the summary methods.
 See the \link{post.processing} method for more details about it.
}
\description{
Landmarking measures are simple and fast algorithms, from which performance
characteristics can be extracted. The measures use k-fold cross-validation
and the evaluation measure is accuracy. For multi-class classification
problems, a decomposition strategy is applied.
}
\details{
The following features are allowed for this method:
 \describe{
   \item{"decision.stumps"}{Construct a single DT node model induced by the
     most informative attribute. The single split (parallel axis) in the data
     has the main goal of establish the linear separability.}
   \item{"elite.nearest.neighbor"}{Select the most informative attributes in
     the dataset using the information gain ratio to induce the 1-Nearest
     Neighbor. With the subset of informative attributes is expected that the
     models induced by 1-Nearest Neighbor should be noise tolerant.}
   \item{"linear.discriminant"}{Apply the Linear Discriminant classifier to
     construct a linear split (non parallel axis) in the data to establish
     the linear separability.}
   \item{"naive.bayes"}{Evaluate the performance of the Naive Bayes
     classifier. It assumes that the attributes are independent and each
     example belongs to a certain class based on the Bayes probability.}
   \item{"nearest.neighbor"}{This measure evaluate the performance of the
     1-Nearest Neighbor classifier. It uses the euclidean distance of the
     nearest neighbor to determine how noisy is the data.}
   \item{"worst.node"}{Construct a single DT node model induced by the
     less informative attribute. With the "decision.stumps" measure is
     possible to define a baseline value of linear separability for a
     dataset.}
 }
}
\examples{
## Extract all meta-features using formula
mf.landmarking(Species ~ ., iris)

## Extract all meta-features using data.frame
mf.landmarking(iris[1:4], iris[5])

## Extract some meta-features
mf.landmarking(Species ~ ., iris, features=c("decision.stumps",
"nearest.neighbor", "linear.discriminant"))

## Extract all meta-features with different summary methods
mf.landmarking(Species ~ ., iris, summary=c("min", "median", "max"))

## Extract all meta-features one vs one decomposition strategy
mf.landmarking(Species ~ ., iris, map="one.vs.one")
}
\references{
Pfahringer, B., Bensusan, H., &  Giraud-Carrier, C. G. (2000). Meta-Learning
 by Landmarking Various Learning Algorithms. In Proceedings of the 17th
 International Conference on Machine Learning (pp. 743-750)
}
\seealso{
Other meta-features: \code{\link{mf.discriminant}},
  \code{\link{mf.general}}, \code{\link{mf.infotheo}},
  \code{\link{mf.model.based}},
  \code{\link{mf.statistical}}
}
